Move precompile to test-setup
