+++ jest.config.js
@@ -1,4 +1,6 @@
 module.exports = {
   ...require('./node_modules/@connectedcars/setup/jest.config.js'),
-  roots: ['<rootDir>/src', '<rootDir>/bin']
+  roots: ['<rootDir>/src', '<rootDir>/bin'],
+  globalSetup: './src/test-setup.ts',
+  globalTeardown: './src/test-teardown.ts'
 }
