+++ package.json
@@ -4,6 +4,9 @@
   "description": "Test utilities for Node projects",
   "main": "build/dist/src/index.js",
   "types": "build/dist/index.d.ts",
+  "bin": {
+    "run-wrapper": "build/dist/bin/run-wrapper.js"
+  },
   "scripts": {
     "test": "TZ=UTC jest",
     "build": "build src bin",
